Consolidate `(when)` option handling
