+++ context/src/main/proto/spine/validation/events.proto
@@ -36,9 +36,7 @@ option java_outer_classname = "EventsProto";
 option java_multiple_files = true;
 
 import "spine/compiler/ast.proto";
-import "spine/validation/time_field_type.proto";
 import "spine/validation/field_group.proto";
-import "spine/time_options.proto";
 
 // A required field has been discovered.
 //
@@ -160,25 +158,6 @@ message IfSetAgainOptionDiscovered {
 
     // The custom error message template.
     string custom_error_message = 3;
-}
-
-// The event emitted whenever a field with `(when)` option is discovered
-// and has passed the necessary checks to confirm the option is applied correctly.
-message WhenFieldDiscovered {
-
-    compiler.FieldRef id = 1;
-
-    // The field in which the option was discovered.
-    compiler.Field subject = 2;
-
-    // The error message template.
-    string error_message = 3;
-
-    // Defines a restriction for the timestamp.
-    Time bound = 4;
-
-    // The type of the field.
-    spine.validation.TimeFieldType type = 5;
 }
 
 // The event emitted whenever a `oneof` group with `(choice).required = true` option is
