+++ context/src/main/proto/spine/validation/views.proto
@@ -36,8 +36,6 @@ option java_outer_classname = "ViewsProto";
 option java_multiple_files = true;
 
 import "spine/compiler/ast.proto";
-import "spine/time_options.proto";
-import "spine/validation/time_field_type.proto";
 import "spine/validation/field_group.proto";
 
 // A view on a field that is recognized as required.
@@ -128,25 +126,6 @@ message DistinctField {
 
     // The error message template.
     string error_message = 3;
-}
-
-// A view of a field that is marked with `(when)` option.
-message WhenField {
-    option (entity).kind = PROJECTION;
-
-    compiler.FieldRef id = 1;
-
-    // The field in which the option was discovered.
-    compiler.Field subject = 2;
-
-    // The error message template.
-    string error_message = 3;
-
-    // Defines a restriction for the timestamp.
-    Time bound = 4;
-
-    // The type of the field.
-    spine.validation.TimeFieldType type = 5;
 }
 
 // A view of a `oneof` group that is marked with `(choice).required = true` option.
