+++ java/src/main/proto/spine/time/validation/events.proto
@@ -1,11 +1,11 @@
 /*
- * Copyright 2025, TeamDev. All rights reserved.
+ * Copyright 2026, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
@@ -26,24 +26,34 @@
 
 syntax = "proto3";
 
-package spine.validation;
+package spine.time.validation;
 
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.spine.io";
-option java_package = "io.spine.tools.validation";
-option java_outer_classname = "TimeFieldTypeProto";
+option java_package = "io.spine.time.validation.event";
+option java_outer_classname = "EventsProto";
 option java_multiple_files = true;
 
-// Time-related field type supported by the `(when)` option.
-enum TimeFieldType {
+import "spine/compiler/ast.proto";
+import "spine/time_options.proto";
+import "spine/time/validation/time_field_type.proto";
 
-    TFT_UNKNOWN = 0;
+// The event emitted whenever a field with `(when)` option is discovered
+// and has passed the necessary checks to confirm the option is applied correctly.
+message WhenFieldDiscovered {
 
-    // Denotes `com.google.protobuf.Timestamp`.
-    TFT_TIMESTAMP = 1;
+    compiler.FieldRef id = 1;
 
-    // Denotes an interface `io.spine.time.Temporal`, which the field type should
-    // implement to be handled by the option.
-    TFT_TEMPORAL = 2;
+    // The field in which the option was discovered.
+    compiler.Field subject = 2;
+
+    // The error message template.
+    string error_message = 3;
+
+    // Defines a restriction for the timestamp.
+    Time bound = 4;
+
+    // The type of the field.
+    spine.time.validation.TimeFieldType type = 5;
 }
