+++ java/src/main/proto/spine/time/validation/time_field_type.proto
@@ -1,11 +1,11 @@
 /*
- * Copyright 2026, TeamDev. All rights reserved.
+ * Copyright 2025, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * https://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
@@ -24,44 +24,26 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import io.spine.dependency.lib.AutoService
-import io.spine.dependency.lib.AutoServiceKsp
-import io.spine.dependency.local.Compiler
+syntax = "proto3";
 
-buildscript {
-    standardSpineSdkRepositories()
-    configurations {
-        all {
-            resolutionStrategy {
-                force(
-                    io.spine.dependency.local.Logging.grpcContext,
-                    io.spine.dependency.lib.JetBrainsAnnotations.lib,
-                )
-            }
-        }
-    }
-    dependencies {
-        classpath(io.spine.dependency.local.CoreJvmCompiler.pluginLib)
-    }
-}
+package spine.time.validation;
 
-plugins {
-    module
-    id("com.google.devtools.ksp")
-}
+import "spine/options.proto";
 
-apply(plugin = "io.spine.core-jvm")
+option (type_url_prefix) = "type.spine.io";
+option java_package = "io.spine.time.validation";
+option java_outer_classname = "TimeFieldTypeProto";
+option java_multiple_files = true;
 
-dependencies {
-    api(Compiler.jvm)
-    api(project(":context"))
-    api(project(":jvm-runtime"))
-    ksp(AutoServiceKsp.processor)
-    compileOnly(AutoService.annotations)
-}
+// Time-related field type supported by the `(when)` option.
+enum TimeFieldType {
 
-afterEvaluate {
-    val kspKotlin by tasks.getting
-    val launchSpineCompiler by tasks.getting
-    kspKotlin.dependsOn(launchSpineCompiler)
+    TFT_UNKNOWN = 0;
+
+    // Denotes `com.google.protobuf.Timestamp`.
+    TFT_TIMESTAMP = 1;
+
+    // Denotes an interface `io.spine.time.Temporal`, which the field type should
+    // implement to be handled by the option.
+    TFT_TEMPORAL = 2;
 }
